Do not set type hint and return type "mixed" on PHP < 8
Mixed is not a valid type on PHP < 8, it was introduced in 8.0.0 by https://wiki.php.net/rfc/mixed_type_v2
When useReturnType (getter) and/or useTypeHints (setter) are enabled, the assembler sets "mixed" both in docblocks and as PHP types in the code. The latter causes issues on PHP < 8.